chore: move creationGuide hook
