+++ apps/calendar/src/components/daygrid/dayGrid.tsx
@@ -19,6 +19,7 @@ import { toEndOfDay, toStartOfDay } from '@src/time/datetime';
 import { PopupType } from '@src/modules/layerPopup';
 import { GridGuideInfo } from '@t/components/daygrid/creationGuide';
 import { GridGuideCreationInfo } from '@t/components/daygrid/gridWithMouse';
+import { useCreationGuide } from '@src/components/hooks/creationGuide';
 
 const TOTAL_PERCENT_HEIGHT = 100;
 
@@ -40,68 +41,6 @@ function useGridHeight() {
   }, []);
 
   return { ref, height };
-}
-
-function useCreationGuide(useCreationPopup = false) {
-  const [creationGuide, setCreationGuide] = useState<GridGuideCreationInfo | null>(null);
-  const [popupFlag, setPopupFlag] = useState(false);
-
-  const { show, hide } = useActions('layerPopup');
-
-  const onOpenCreationPopup = (guide: GridGuideCreationInfo) => {
-    if (useCreationPopup) {
-      const { start, end } = guide;
-
-      // @TODO: popupRect 계산 필요
-      show({
-        type: PopupType.creation,
-        param: {
-          start,
-          end,
-          isAllDay: true,
-          popupRect: {
-            width: 474,
-            height: 272,
-            left: 102.695,
-            top: 257,
-          },
-          close: () => {
-            onGuideCancel();
-            setPopupFlag(false);
-          },
-        },
-      });
-
-      if (!popupFlag) {
-        setPopupFlag(true);
-      }
-    }
-  };
-
-  const onGuideStart = (guide: GridGuideCreationInfo | null) => {
-    setCreationGuide(guide);
-  };
-  const onGuideEnd = (guide: GridGuideCreationInfo | null) => {
-    setCreationGuide(guide);
-
-    if (guide) {
-      onOpenCreationPopup(guide);
-    } else if (!guide && popupFlag) {
-      hide();
-    }
-  };
-  const onGuideChange = (guide: GridGuideCreationInfo) => {
-    setCreationGuide(guide);
-  };
-  const onGuideCancel = () => setCreationGuide(null);
-
-  return {
-    creationGuide,
-    onGuideStart,
-    onGuideEnd,
-    onGuideChange,
-    onGuideCancel,
-  };
 }
 
 function renderCreationGuide(
